+++ QueenBeeParser/Zlib/Tree.cs
@@ -187,7 +187,7 @@ namespace Rebex.IO.Compression
 						continue;
 					if (tree[m * 2 + 1] != bits)
 					{
-						s.opt_len = (int) (s.opt_len + ((long) bits - (long) tree[m * 2 + 1]) * (long) tree[m * 2]);
+						s.opt_len = (int) (s.opt_len + (bits - (long) tree[m * 2 + 1]) * tree[m * 2]);
 						tree[m * 2 + 1] = (short) bits;
 					}
 					n--;
@@ -268,7 +268,7 @@ namespace Rebex.IO.Compression
 				
 				// Create a new node father of n and m
 				tree[node * 2] = (short) (tree[n * 2] + tree[m * 2]);
-				s.depth[node] = (byte) (System.Math.Max((byte) s.depth[n], (byte) s.depth[m]) + 1);
+				s.depth[node] = (byte) (System.Math.Max(s.depth[n], s.depth[m]) + 1);
 				tree[n * 2 + 1] = tree[m * 2 + 1] = (short) node;
 				
 				// and insert the new node in the heap
